Move DisconnectErrors and INebulaConnection to `NebulaAPI.Networking` -
namespace and update necessary references
