+++ NebulaModel/Packets/PacketProcessor.cs
@@ -1,5 +1,6 @@
 ﻿#region
 
+using NebulaAPI.Networking;
 using NebulaAPI.Packets;
 using NebulaModel.Networking;
